переименовал компоненты
