+++ src/components/movie-card/movie-card.jsx
@@ -1,31 +1,20 @@
-
 import React from 'react';
 import PropTypes from 'prop-types';
-import Header from '../header/header';
-import MovieCardInfo from '../movie-card-info/movie-card-info';
 
-const MovieCard = ({background_image: background, name, ...rest}) => (
-  <section className="movie-card">
-    <div className="movie-card__bg">
-      <img src={background} alt={name} />
+const SmallMovieCard = ({name, preview}) => (
+  <article className="small-movie-card catalog__movies-card">
+    <div className="small-movie-card__image">
+      <img src={preview} alt={name} width="280" height="175" />
     </div>
-
-    <h1 className="visually-hidden">WTW</h1>
-
-    <Header />
-
-    <div className="movie-card__wrap">
-      <MovieCardInfo
-        name = {name}
-        {...rest}
-      />
-    </div>
-  </section>
+    <h3 className="small-movie-card__title">
+      <a className="small-movie-card__link" href="movie-page.html">{name}</a>
+    </h3>
+  </article>
 );
 
-MovieCard.propTypes = {
-  "background_image": PropTypes.string.isRequired,
-  "name": PropTypes.string.isRequired
+SmallMovieCard.propTypes = {
+  preview: PropTypes.string.isRequired,
+  name: PropTypes.string.isRequired
 };
 
-export default MovieCard;
+export default SmallMovieCard;
